+++ simter-file-dao-jpa/src/main/kotlin/tech/simter/file/impl/dao/jpa/FileBlockDao.kt
@@ -23,7 +23,7 @@ interface FileBlockDao {
     moduleMatcher: ModuleMatcher,
     search: Optional<String> = Optional.empty(),
     limit: Int,
-    offset: Int = 0
+    offset: Long = 0
   ): Page<FileStore>
 
   /**
